Adding parsers for wavefunctions and U matrices
A parser for the _u.mat and _u_dis.mat files has been added.
Moreover, also a parser for wavefunctions (UNKNNNNN.N files)
has been added, but also in the (non-default) case where they
are printed in formatted format, for now.

Tests are added (skipping U matrices for Wannier 2.x where this
was not implemented) and CLI is expanded as well with two new commands.